requirements.txt
Django==2.2.3
django-debug-toolbar==2.0
django-fontawesome==1.0
django-phone-field==1.7.1
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.1
PyYAML==5.1.2
sqlparse==0.3.0
